+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.49.0] - 2018-08-23
 ### Added
 - CLI: Added openapi3 schema generation option to update openapi3 YAML file definition with schemas.